+++ geotrek/altimetry/tests/test_commands.py
@@ -99,8 +99,8 @@ class CommandLoadDemTest(TransactionTestCase):
             geom=LineString((605600, 6650000), (605900, 6650010), srid=2154)
         )
         with self.assertNumQueries(
-            22
-        ):  # 5 for loaddem initial + 17 with selects and update geom
+            23
+        ):  # 5 for loaddem initial + 17 with selects and update geom + 1 for PointTopologyTestModel
             call_command(
                 "loaddem",
                 filename,