+++ geotrek/infrastructure/parsers.py
@@ -1,10 +1,11 @@
-from django.conf import settings
-from django.contrib.gis.geos import Point
+from django.contrib.gis.geos import GEOSGeometry, Point
+from django.core.exceptions import ImproperlyConfigured
 from django.utils.translation import gettext as _
 
-from geotrek.common.parsers import GeotrekParser, GlobalImportError, OpenStreetMapParser
-from geotrek.core.models import Path, Topology
+from geotrek.common.parsers import GeotrekParser, OpenStreetMapParser
+from geotrek.core.mixins.parsers import PointTopologyParserMixin
 from geotrek.infrastructure.models import Infrastructure
+from geotrek.trekking.parsers import ApidaeBaseParser
 
 
 class GeotrekInfrastructureParser(GeotrekParser):
@@ -32,7 +33,62 @@ class GeotrekInfrastructureParser(GeotrekParser):
         self.next_url = f"{self.url}/api/v2/infrastructure"
 
 
-class OpenStreetMapInfrastructureParser(OpenStreetMapParser):
+class ApidaeInfrastructureParser(PointTopologyParserMixin, ApidaeBaseParser):
+    model = Infrastructure
+    eid = "eid"
+    separator = None
+    infrastructure_type = None
+
+    responseFields = [
+        "id",
+        "localisation.geolocalisation.geoJson",
+        "nom",
+        "presentation",
+        "prestations.tourismesAdaptes",
+    ]
+    fields = {
+        "eid": "id",
+        "geom": "localisation.geolocalisation.geoJson",
+        "name": "nom.libelleFr",
+        "description": (
+            "presentation.descriptifCourt.libelleFr",
+            "presentation.descriptifDetaille.libelleFr",
+        ),
+        "accessibility": "prestations.tourismesAdaptes",
+    }
+    natural_keys = {
+        "type": "label",
+    }
+    field_options = {
+        "type": {"create": True},
+        "name": {"required": True},
+        "geom": {"required": True},
+    }
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        if self.infrastructure_type:
+            self.constant_fields = self.constant_fields.copy()
+            self.constant_fields["type"] = self.infrastructure_type
+        else:
+            msg = _(
+                "An infrastructure type must be specified in the parser configuration."
+            )
+            raise ImproperlyConfigured(msg)
+
+    def build_geos_geometry(self, src, val):
+        return GEOSGeometry(str(val))
+
+    def filter_description(self, src, val):
+        short_descr, detailed_descr = val
+        return detailed_descr or short_descr
+
+    def filter_accessibility(self, src, val):
+        accessibilities = [a.get("libelleFr") for a in val]
+        return "\n".join(accessibilities)
+
+
+class OpenStreetMapInfrastructureParser(PointTopologyParserMixin, OpenStreetMapParser):
     """Parser to import infrastructures from OpenStreetMap"""
 
     type = None
@@ -53,15 +109,7 @@ class OpenStreetMapInfrastructureParser(OpenStreetMapParser):
         if self.type:
             self.constant_fields["type"] = self.type
 
-    def start(self):
-        super().start()
-        if settings.TREKKING_TOPOLOGY_ENABLED and not Path.objects.exists():
-            raise GlobalImportError(
-                _("You need to add a network of paths before importing Infrastructures")
-            )
-
-    def filter_geom(self, src, val):
-        # convert OSM geometry to point
+    def build_geos_geometry(self, src, val):
         type, lng, lat, area, bbox = val
         geom = None
         if type == "node":
@@ -70,20 +118,4 @@ class OpenStreetMapInfrastructureParser(OpenStreetMapParser):
             geom = self.get_centroid_from_way(area)
         elif type == "relation":
             geom = self.get_centroid_from_relation(bbox)
-
-        # create topology
-        self.topology = Topology.objects.none()
-        if settings.TREKKING_TOPOLOGY_ENABLED:
-            # Use existing topology helpers to transform a Point(x, y)
-            # to a path aggregation (topology)
-            serialized = f'{{"lng": {geom.x}, "lat": {geom.y}}}'
-            self.topology = Topology.deserialize(serialized)
-            # Move deserialization aggregations to the POI
-
-        geom.transform(settings.SRID)
         return geom
-
-    def parse_obj(self, row, operation):
-        super().parse_obj(row, operation)
-        if settings.TREKKING_TOPOLOGY_ENABLED and self.obj.geom and self.topology:
-            self.obj.mutate(self.topology)